map working
map added to GeoChart page
