+++ src/components/title.jsx
@@ -1,0 +1,5 @@
+export default function Title() {
+  return (
+    <h1>World map using Geo Charts</h1>
+  );
+}
