--- src/title.jsx
@@ -1,8 +1,0 @@
-import './styles.css';
-import Map from './map';
-
-export default function Title() {
-  return (
-    <h1>F1 Around The World</h1>
-  );
-}
